add dlc working

+++ gamemanager/Models/GameEntry.cs
@@ -18,12 +18,12 @@ namespace gamemanager.Models
     public class Dlc
     {
         public int Id { get; set; }
-        public int ParentGameId { get; set; }
+        public long ParentGameId { get; set; }
         public string Name { get; set; }
         public bool Owned { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Notes { get; set; }
-        public int Rating { get; set; }
-        public int Ranking { get; set; }
+        public short Rating { get; set; }
+        public short Ranking { get; set; }
     }
 }
